// src/api/generated/model/authenticationResponse.ts
/**
 * Generated by orval v8.0.1 🍺
 * Do not edit manually.
 * Measuring Parent API
 * API documentation for Measuring Parent application
 * OpenAPI spec version: 1.0.0
 */
import type { UserDto } from "./userDto";

export interface AuthenticationResponse {
  accessToken?: string;
  refreshToken?: string;
  user?: UserDto;
}