// src/api/generated/model/authenticationLearnerResponse.ts
/**
 * Generated by orval v8.0.1 🍺
 * Do not edit manually.
 * Measuring Parent API
 * API documentation for Measuring Parent application
 * OpenAPI spec version: 1.0.0
 */
import type { ApplicationDto } from "./applicationDto";
import type { CandidateDto } from "./candidateDto";
import type { EvaluationDto } from "./evaluationDto";
import type { ExamDto } from "./examDto";
import type { ExamSessionDto } from "./examSessionDto";

export interface AuthenticationLearnerResponse {
  accessToken?: string;
  refreshToken?: string;
  user?: CandidateDto;
  examSession?: ExamSessionDto;
  exam?: ExamDto;
  application?: ApplicationDto;
  evaluations?: EvaluationDto[];
}